// src/cliente/dto/create-cliente.dto.ts
export class CreateClienteDto {
  usuario: string;
  correoElectronico: string;
  nombre: string;
  apellidos: string;
  edad: string;
  genero: string;
  numTelefono: string;
  rfc: string;
  fechaNacimiento: string;
  calleNum: string;
  colonia: string;
  cp: string;
  municipio: string;
  metodoPago: string;
}